src/requirements.txt
matplotlib
numpy
Pillow
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
opencv-python
